Organize files into structure folder
