+++ src/worldgen/index.ts
@@ -9,12 +9,7 @@ export * from './NoiseGeneratorSettings.js'
 export * from './NoiseRouter.js'
 export * from './NoiseSettings.js'
 export * from './RandomState.js'
-export * from './StructurePlacement.js'
-export * from './StructurePoolElement.js'
-export * from './StructureSet.js'
-export * from './StructureTemplatePool.js'
+export * from './structure/index.js'
 export * from './SurfaceSystem.js'
 export * from './VerticalAnchor.js'
 export * from './WorldgenRegistries.js'
-export * from './WorldgenStructure.js'
-
