+++ src/worldgen/structure/WorldgenStructure.ts
@@ -1,17 +1,17 @@
-import { BlockPos, Holder, HolderSet, Identifier, Registry } from '../core/index.js'
-import { Rotation } from '../core/Rotation.js'
-import type { Random } from '../math/index.js'
-import { LegacyRandom } from '../math/index.js'
-import { Json } from '../util/Json.js'
-import type { BiomeSource } from './biome/index.js'
-import { Heightmap } from './Heightmap.js'
-import { HeightProvider } from './HeightProvider.js'
-import { NoiseChunkGenerator } from './NoiseChunkGenerator.js'
-import type { NoiseGeneratorSettings } from './NoiseGeneratorSettings.js'
-import { RandomState } from './RandomState.js'
+import { Rotation } from '../../core/Rotation.js'
+import { BlockPos, Holder, HolderSet, Identifier, Registry } from '../../core/index.js'
+import type { Random } from '../../math/index.js'
+import { LegacyRandom } from '../../math/index.js'
+import { Json } from '../../util/Json.js'
+import { HeightProvider } from '../HeightProvider.js'
+import { Heightmap } from '../Heightmap.js'
+import { NoiseChunkGenerator } from '../NoiseChunkGenerator.js'
+import type { NoiseGeneratorSettings } from '../NoiseGeneratorSettings.js'
+import { RandomState } from '../RandomState.js'
+import { WorldgenRegistries } from '../WorldgenRegistries.js'
+import type { BiomeSource } from '../biome/index.js'
 import { StructurePoolElement } from './StructurePoolElement.js'
 import { StructureTemplatePool } from './StructureTemplatePool.js'
-import { WorldgenRegistries } from './WorldgenRegistries.js'
 
 export type SufaceLevelAccessor = (posX: number, posZ: number, heightmap: Heightmap) => number
