+++ src/worldgen/structure/index.ts
@@ -1,0 +1,5 @@
+export * from './StructurePlacement.js'
+export * from './StructurePoolElement.js'
+export * from './StructureSet.js'
+export * from './StructureTemplatePool.js'
+export * from './WorldgenStructure.js'
